Export CopyToClipboard from main entry point
